feat(admin): New Layer Page UI [EP-2200]
