+++ package.json
@@ -53,7 +53,6 @@
     "react-slick": "^0.23.2",
     "react-spring": "8.0.4",
     "react-select": "^3.1.0",
-    "react-select-fetch": "^0.2.0",
     "react-select-async-paginate": "^0.4.0",
     "recharts": "^1.5.0",
     "redux": "^4.0.1",
