+++ packages/earth-admin/src/pages-client/layers/model.ts
@@ -81,3 +81,32 @@ export enum LayerCategory {
   SOCIO_ECONOMIC = 'Socio-Economic',
 }
 
+
+export const LAYER_CATEGORY_OPTIONS = [
+  { value: 'Biodiversity', label: 'Biodiversity' },
+  { value: 'Climate & Carbon', label: 'Climate & Carbon' },
+  { value: 'Ecosystem Services', label: 'Ecosystem Services' },
+  { value: 'Human Impact', label: 'Human Impact' },
+  { value: 'Land Cover', label: 'Land Cover' },
+  { value: 'Marine', label: 'Marine' },
+  { value: 'Natural Hazards', label: 'Natural Hazards' },
+  { value: 'Protected Areas', label: 'Protected Areas' },
+  { value: 'Restoration', label: 'Restoration' },
+  { value: 'Socio-Economic', label: 'Socio-Economic' }
+]
+
+
+export const LAYER_TYPE_OPTIONS = [
+  { value: 'raster', label: 'raster' },
+  { value: 'vector', label: 'vector' },
+  { value: 'geojson', label: 'geojson' },
+  { value: 'group', label: 'group' },
+  { value: 'video', label: 'video' }
+]
+
+export const LAYER_PROVIDER_OPTIONS = [
+  { value: 'cartodb', label: 'cartodb' },
+  { value: 'gee', label: 'gee' },
+  { value: 'mapbox', label: 'mapbox' },
+  { value: 'leaflet', label: 'leaflet' }
+]
